adding ui

+++ app/src/main/java/com/example/task_five_sqlite3/ContactDbHelper.kt
@@ -74,7 +74,7 @@ class ContactDbHelper(context: Context) :
         return id
     }
 
-    fun deleteContact(dbHelper: ContactDbHelper, id: Long) {
+    fun deleteContact(dbHelper: ContactDbHelper, id: Int) {
         val db = dbHelper.writableDatabase
         db.delete(ContactContract.ContactEntry.TABLE_NAME,
             "${BaseColumns._ID} = ?",
